+++ docs/handbook/sources/build.sh
@@ -272,6 +272,7 @@ sed '1,19d' ../../../library/redis/NOTES.md | pandoc -f gfm -t rst -o libraries/
 sed '1,19d' ../../../library/regression_protocols/NOTES.md | pandoc -f gfm -t rst -o libraries/regression_protocols.rst
 sed '1,19d' ../../../library/regression_tree/NOTES.md | pandoc -f gfm -t rst -o libraries/regression_tree.rst
 sed '1,19d' ../../../library/regularized_bradley_terry_ranker/NOTES.md | pandoc -f gfm -t rst -o libraries/regularized_bradley_terry_ranker.rst
+sed '1,19d' ../../../library/rest/NOTES.md | pandoc -f gfm -t rst -o libraries/rest.rst
 sed '1,19d' ../../../library/ridge_regression/NOTES.md | pandoc -f gfm -t rst -o libraries/ridge_regression.rst
 sed '1,19d' ../../../library/schulze_ranker/NOTES.md | pandoc -f gfm -t rst -o libraries/schulze_ranker.rst
 sed '1,19d' ../../../library/sequential_pattern_mining_protocols/NOTES.md | pandoc -f gfm -t rst -o libraries/sequential_pattern_mining_protocols.rst