+++ paths/paths_examples.pl
@@ -108,6 +108,8 @@ logtalk_library_path(hooks, examples('hooks/')).
 logtalk_library_path(http_cookies_counter, examples('http_cookies_counter/')).
 logtalk_library_path(http_htmx_panel, examples('http_htmx_panel/')).
 logtalk_library_path(http_open_api, examples('http_open_api/')).
+logtalk_library_path(http_rest_greetings, examples('http_rest_greetings/')).
+logtalk_library_path(http_rest_open_meteo, examples('http_rest_open_meteo/')).
 logtalk_library_path(http_static_site, examples('http_static_site/')).
 logtalk_library_path(http_websocket_echo, examples('http_websocket_echo/')).
 logtalk_library_path(includes, examples('includes/')).